Scripts: NFS + NTP

+++ scripts/nfs/nfs_server-conf.sh
@@ -4,7 +4,9 @@
 apt-get install nfs-kernel-server nfs-common
 
 # Create a backup of the configuration file.
-cp /etc/exports /etc/exports.bk
+if [ ! -f /etc/exports.bk ]; then
+    cp /etc/exports /etc/exports.bk
+fi
 
 # Create the sharing directory.
 mkdir /srv/nfs
@@ -13,7 +15,10 @@ mkdir /srv/nfs
 echo "###### NFS SERVER CONFIGURATION ######" > /etc/exports
 echo " " >> /etc/exports
 
-echo "/srv/nfs	10.1.0.0/16(rw,root_squash)"
+echo "/srv/nfs	10.1.0.0/16(rw,root_squash,no_subtree_check)" >> /etc/exportfs
+
+# Update the table of exported file systems.
+exports -a
 
 # Restart NFS.
 /etc/init.d/nfs-kernel-server reload
